Add version into devices.json.

+++ src/main/java/com/hbm/devices/scan/ui/android/DeviceZipper.java
@@ -77,12 +77,14 @@ final class DeviceZipper {
                 }
             }
             File file = new File(subDir, FILE_NAME);
-            if (!file.delete()) {
-                Toast.makeText(activity, activity.getString(R.string.could_not_delete) + FILE_NAME, Toast.LENGTH_SHORT).show();
-                return null;
+            if (file.exists()) {
+                if (!file.delete()) {
+                    Toast.makeText(activity, activity.getString(R.string.could_not_delete, FILE_NAME), Toast.LENGTH_SHORT).show();
+                    return null;
+                }
             }
             if (!file.createNewFile()) {
-                Toast.makeText(activity, activity.getString(R.string.could_not_create) + FILE_NAME, Toast.LENGTH_SHORT).show();
+                Toast.makeText(activity, activity.getString(R.string.could_not_create, FILE_NAME), Toast.LENGTH_SHORT).show();
                 return null;
             }
             FileOutputStream fos = new FileOutputStream(file, false);
@@ -90,6 +92,7 @@ final class DeviceZipper {
             ZipEntry entry = new ZipEntry("devices.json");
             zos.putNextEntry(entry);
             zos.write(("{\"date\":\"" + isoDate + "\",").getBytes(charSet));
+            zos.write(("\"version\": \"1.0\",").getBytes(charSet));
             zos.write("\"devices\":[".getBytes(charSet));
 
             Iterator<Announce> iterator = announces.iterator();
